\docType{package}
\name{econresearchR}
\alias{econresearchR}
\title{econresearchR: Organizing files in applied economic research}
\description{
This is a useful package for organizing files in applied economic research.
}
\details{
\itemize{
\item{Package: econresearchR}
\item{Version: 1.0}
\item{Maintainer: Ronny M. Condor \email{ronny.condor@unmsm.edu.pe}}
}
}

\author{
\itemize{
\item{Condor Iturrizaga, Ronny Martin}
}
}

